+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # rcc change log
 
+## v11.0.7 (date: 14.9.2021) UNSTABLE
+
+- BREAKING CHANGES (ongoing work, see v11.0.0 for more details)
+- changed progress indication to match holotree flow
+- made log and telemetry waiting visible in timeline
+
 ## v11.0.6 (date: 13.9.2021) UNSTABLE
 
 - BREAKING CHANGES (ongoing work, see v11.0.0 for more details)